refactor(account): rename vpc -> cloud-native-network; rename --allow-create-new
- Rename 'account vpc' subcommand group to 'account cloud-native-network'
  to match the platform's terminology (the underlying resources are
  AccountConfig CloudNativeNetworks, not just AWS VPCs).
- Rename --vpc-id/--vpc-ids flags to --network-id/--network-ids.
- Rename --allow-create-new to --allow-create-new-cloud-native-network so
  it's obvious what is being created (matches the BYOA injected input
  parameter 'Allow New Cloud Native Network Creation').
- Rename internal package cmd/account/vpc -> cmd/account/cloudnativenetwork
  and dataaccess functions *AccountConfigVPC* -> *AccountConfigCloudNativeNetwork*.
- Regenerate CLI docs.
