+++ cmd/account/account.go
@@ -1,7 +1,7 @@
 package account
 
 import (
-	"github.com/omnistrate-oss/omnistrate-ctl/cmd/account/vpc"
+	"github.com/omnistrate-oss/omnistrate-ctl/cmd/account/cloudnativenetwork"
 	"github.com/spf13/cobra"
 )
 
@@ -20,7 +20,7 @@ func init() {
 	Cmd.AddCommand(listCmd)
 	Cmd.AddCommand(describeCmd)
 	Cmd.AddCommand(deleteCmd)
-	Cmd.AddCommand(vpc.Cmd)
+	Cmd.AddCommand(cloudnativenetwork.Cmd)
 }
 
 func run(cmd *cobra.Command, args []string) {
