+++ mkdocs/docs/omnistrate-ctl_account.md
@@ -26,11 +26,11 @@ omnistrate-ctl account [operation] [flags]
 ### SEE ALSO
 
 * [omnistrate-ctl](omnistrate-ctl.md)	 - Manage your Omnistrate SaaS from the command line
+* [omnistrate-ctl account cloud-native-network](omnistrate-ctl_account_cloud-native-network.md)	 - Manage cloud-native networks (VPCs) for a BYOA Cloud Provider Account
 * [omnistrate-ctl account create](omnistrate-ctl_account_create.md)	 - Create a Cloud Provider Account
 * [omnistrate-ctl account customer](omnistrate-ctl_account_customer.md)	 - Manage customer BYOA account onboarding
 * [omnistrate-ctl account delete](omnistrate-ctl_account_delete.md)	 - Delete a Cloud Provider Account
 * [omnistrate-ctl account describe](omnistrate-ctl_account_describe.md)	 - Describe a Cloud Provider Account
 * [omnistrate-ctl account list](omnistrate-ctl_account_list.md)	 - List Cloud Provider Accounts
 * [omnistrate-ctl account update](omnistrate-ctl_account_update.md)	 - Update a Cloud Provider Account
-* [omnistrate-ctl account vpc](omnistrate-ctl_account_vpc.md)	 - Manage VPCs for a Cloud Provider Account
